Rework gopls start/stop/restart
This adds two extension preferences:

* enable/disable gopls (default disabled)
* path to gopls (use a gopls found in the search path)

The start/stop/restart is all new, based on Language Server Extension template from Panic.

There is a hack around what seems like a LanguageClient bug, in that there is not a straight forward way to establish when, after calling stop(), the LanguageClient is fully cleaned up. The `running` property persists in reporting the client to be running. This becomes problematic in restarts because creating a new client fails if the old client hasn’t finished shutting down. For now, the restart scenario just waits a bit between stopping and starting.